# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Simva, AKA Simple Validator, AKA SurveyManager is the evolution of a tool created to integrate and ease the task of validating serious games using pre-post surveys as a control point. This version allows to create more complex studies with more variety of activities and different allocations.",
    "version": "1.0.0",
    "title": "Simva API",
    "contact": {
      "email": "ivanjper@ucm.es"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/limesurvey-completion-webhooks": {
      "post": {
        "summary": "get completion from limesurvey",
        "operationId": "GetEvent",
        "description": "get completion from limesurvey\n",
        "responses": {
          "200": {
            "description": "completion from limesurvey"
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "401": {
            "description": "One or both of the JWT are not valid"
          },
          "required": true
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "admins"
        ],
        "summary": "gets the list of users.",
        "operationId": "getUsers",
        "description": "Obtains the list of groups owned by current user.\n",
        "responses": {
          "200": {
            "description": "the list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/user"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/searchString"
          },
          {
            "$ref": "#/components/parameters/skip"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ]
      },
      "post": {
        "summary": "adds a user.",
        "operationId": "addUser",
        "description": "Creates a new group for the current user as owner.\n",
        "responses": {
          "200": {
            "description": "item created"
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/user"
              }
            }
          },
          "required": true
        }
      }
    },
...
    "/groups": {
      "get": {
        "tags": [
          "admins",
          "teachers",
          "students"
        ],
        "summary": "gets the list of groups where this student pl",
        "operationId": "getGroups",
        "description": "Obtains the list of groups owned by current user.\n",
        "responses": {
          "200": {
            "description": "the list of groups for this user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/group"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/searchString"
          },
          {
            "$ref": "#/components/parameters/skip"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ]
      },
      "post": {
        "tags": [
          "teachers"
        ],
        "summary": "adds a group for the current user as owner.",
        "operationId": "addGroup",
        "description": "Creates a new group for the current user as owner.\n",
        "responses": {
          "200": {
            "description": "group created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/group"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/add_group_body"
              }
            }
          },
          "required": true
        }
      }
    },
...
    "/studies": {
      "get": {
        "tags": [
          "admins",
          "teachers",
          "students"
        ],
        "summary": "gets the list of studies for the current teacher",
        "operationId": "getStudies",
        "description": "Obtains the list of studies for the current teacher.\n",
        "responses": {
          "200": {
            "description": "the list of studies for this teacher",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/study"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/searchString"
          },
          {
            "$ref": "#/components/parameters/skip"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ]
      },
      "post": {
        "tags": [
          "teachers"
        ],
        "summary": "adds an study for the current teacher",
        "operationId": "addStudy",
        "description": "Creates a new study for the current teacher.\n",
        "responses": {
          "200": {
            "description": "study created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/study"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/add_study_body"
              }
            }
          },
          "description": "Study to be created"
        }
      }
    },
...
    "/activities": {
      "get": {
        "tags": [
          "admins",
          "teachers"
        ],
        "summary": "gets the list of activities for the current teacher",
        "operationId": "getActivities",
        "description": "Obtains the list of activities for the current teacher.\n",
        "responses": {
          "200": {
            "description": "the list of activities for this teacher",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/activity"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/searchString"
          },
          {
            "$ref": "#/components/parameters/skip"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ]
      },
      "post": {
        "tags": [
          "teachers"
        ],
        "summary": "adds a Activity",
        "operationId": "addActivity",
        "description": "Creates a new activity for the current teacher.\n",
        "responses": {
          "200": {
            "description": "activity created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/activity"
                }
              }
            }
          },
          "400": {
            "description": "invalid input, object invalid"
          },
          "409": {
            "description": "an existing item already exists"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/activity"
              }
            }
          },
          "description": "Activity to be created"
        }
      }
    },
...
    "/activitytypes": {
      "get": {
        "tags": [
          "teachers"
        ],
        "summary": "gets the types of activities available",
        "operationId": "getActivityTypes",
        "description": "Obtains the list of activity types available to\nbe created with some additional data\n",
        "responses": {
          "200": {
            "description": "the available activity types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/activity_type"
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/allocatortypes": {
      "get": {
        "tags": [
          "teachers"
        ],
        "summary": "gets the types of allocators available",
        "operationId": "getAllocatorTypes",
        "description": "Obtains the list of allocator types available to\nbe assigned to the studies\n",
        "responses": {
          "200": {
            "description": "the available allocator types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/allocator_type"
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID is not a valid ObjectId"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "OAuth2": [
        "read",
        "write"
      ]
    }
  ],
  "components": {
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://sso.simva.e-ucm.es/auth/realms/simva/protocol/openid-connect/auth",
            "tokenUrl": "https://sso.simva.e-ucm.es/auth/realms/simva/protocol/openid-connect/token",
            "scopes": {
              "read": "Grants read access",
              "write": "Grants write access"
            }
          }
        }
      }
    },
    "parameters": {
...
      "searchString": {
        "in": "query",
        "name": "searchString",
        "description": "pass an optional search string for result filtering",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "skip": {
        "in": "query",
        "name": "skip",
        "description": "number of records to skip for pagination",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 0
        }
      },
      "limit": {
        "in": "query",
        "name": "limit",
        "description": "maximum number of records to return",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 0,
          "maximum": 50
        }
      },
...
    },
...
    "schemas": {
      "user": {
        "type": "object",
        "required": [
          "username",
          "email",
          "password",
          "role"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "username_unique"
          },
          "email": {
            "type": "string",
            "example": "email"
          },
          "password": {
            "type": "string",
            "example": "Pass12345"
          },
          "isToken": {
            "type": "string",
            "example": "True"
          },
          "token": {
            "type": "string",
            "example": "user"
          },
          "role": {
            "type": "string",
            "description": "The role for the user to have. Some\nexamples include teacher, student \nteaching-assistant or researcher\n",
            "enum": [
              "norole",
              "student",
              "teacher",
              "teaching-assistant",
              "researcher",
              "admin"
            ]
          },
          "external_entity": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "domain": {
                  "type": "string",
                  "example": "keycloak"
                },
                "id": {
                  "type": "string",
                  "example": "whateveruuidusedthere_12321"
                }
              }
            }
          }
        }
      },
...
      "group": {
        "type": "object",
        "required": [
          "name",
          "owners",
          "participants",
          "created"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
          },
          "name": {
            "type": "string",
            "example": "Test group for School"
          },
          "owners": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          },
          "participants": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          },
          "link": {
            "type": "object",
            "x-mongoose-type": "Mixed"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "example": "2017-07-21T17:32:28Z"
          },
          "version": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "study": {
        "type": "object",
        "required": [
          "name",
          "tests",
          "allocator",
          "owners",
          "created"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
          },
          "name": {
            "type": "string",
            "example": "Survey for SG validation"
          },
          "tests": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          },
          "allocator": {
            "type": "string",
            "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          },
          "owners": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "example": "2017-07-21T17:32:28Z"
          },
          "active": {
            "type": "string",
            "format": "boolean",
            "example": true
          }
        }
      },
...
      "activity": {
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
          },
          "name": {
            "type": "string",
            "example": "Survey for validation"
          },
          "type": {
            "type": "string",
            "example": "survey, gameplay, etc..."
          },
          "test": {
            "type": "string",
            "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
          },
          "owners": {
            "type": "array",
            "items": {
              "type": "string",
              "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
            }
          },
          "extra_data": {
            "type": "object",
            "properties": {
              "surveyId": {
                "type": "string"
              },
              "surveyOwner": {
                "type": "string"
              },
              "participants": {
                "type": "object",
                "additionalProperties": {
                  "type": "object",
                  "properties": {
                    "completion": {
                      "type": "string",
                      "format": "boolean"
                    },
                    "progress": {
                      "type": "integer"
                    }
                  }
                }
              },
              "config": {
                "type": "object",
                "properties": {
                  "trace_storage": {
                    "type": "string",
                    "format": "boolean"
                  },
                  "realtime": {
                    "type": "string",
                    "format": "boolean"
                  },
                  "backup": {
                    "type": "string",
                    "format": "boolean"
                  }
                }
              },
              "miniotrace": {
                "type": "object",
                "properties": {
                  "presignedUrl": {
                    "type": "string"
                  },
                  "generated_at": {
                    "type": "string"
                  },
                  "expire_on_seconds": {
                    "type": "string"
                  }
                }
              },
              "user_managed": {
                "type": "string",
                "format": "boolean"
              }
            }
          },
          "active": {
            "type": "string",
            "format": "boolean",
            "example": true
          },
          "study": {
            "type": "string",
            "example": "ObjectId(\"5d07a0767c1405004bc702af\")"
          }
        }
      },
...
      "activity_type": {
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "LimeSurvey Activity"
          },
          "type": {
            "type": "string",
            "example": "limesurvey"
          },
          "description": {
            "type": "string",
            "example": "A survey-based activity that uses LimeSurvey as service."
          },
          "utils": {
            "type": "object",
            "x-mongoose-type": "Mixed",
            "additionalProperties": true
          }
        }
      },
      "allocator_type": {
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Group Allocator"
          },
          "type": {
            "type": "string",
            "example": "group"
          },
          "description": {
            "type": "string",
            "example": "This allocator allows you to select to which test branch is each group assigned."
          },
          "utils": {
            "type": "object",
            "x-mongoose-type": "Mixed",
            "additionalProperties": true
          }
        }
      },
...
      "add_group_body": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Name of the group"
          },
          "random_participants": {
            "type": "integer",
            "description": "The amount of codes to be created\nrandomly when the group is created\n",
            "example": 30
          }
        },
        "x-swagger-mongoose": {
          "exclude-schema": true
        }
      },
      "add_study_body": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Name of the Study"
          }
        },
        "x-swagger-mongoose": {
          "exclude-schema": true
        }
      },
...
    }
  }
}